+++ js/util.mjs
@@ -1,0 +1,31 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const createRandomIdFromRange = (min, max) => {
+  const previousValues = [];
+
+  return () => {
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+
+    let currentValue = getRandomInteger(min, max);
+
+    while(previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+
+    previousValues.push(currentValue);
+
+    return currentValue;
+  };
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, createRandomIdFromRange, getRandomArrayElement};
